Added grid and using integers instead of floats

+++ include/Tile.h
@@ -6,18 +6,18 @@
 class Tile : public sf::Sprite
 {
 public:
-	Tile(float const x,
-		float const y,
-		float const width,
-		float const height,
+	Tile(int const x,
+		int const y,
+		int const width,
+		int const height,
 		sf::Texture const& tex);
 
-	sf::Vector2f vec() const;
+	sf::Vector2i vec() const;
 private:
-	float x;
-	float y;
-	float width;
-	float height;
+	int x;
+	int y;
+	int width;
+	int height;
 };
 
 #endif
